+++ packages/collector/test/initEnv.js
@@ -29,26 +29,7 @@ Object.keys(hostsConfig).forEach(key => {
 });
 
 if (!isCI()) {
-  const checksumPath = path.join(__dirname, '.currencies-checksum');
-  const hash = crypto.createHash('md5');
-  hash.update(fs.readFileSync(path.join(rootDir, 'currencies.json'), 'utf8'));
-  hashTemplates(__dirname, hash);
-  const currentHash = hash.digest('hex');
-
-  let needsRegen = true;
-  try {
-    needsRegen = fs.readFileSync(checksumPath, 'utf8').trim() !== currentHash;
-  } catch (_) {
-    // checksum file doesn't exist yet → first run
-  }
-
-  if (needsRegen) {
-    log('[INFO] Test folders out of date — regenerating...');
-    execSync('node bin/create-version-test-folders.js', { cwd: rootDir, stdio: 'inherit' });
-    fs.writeFileSync(checksumPath, currentHash);
-  } else {
-    log('[INFO] Test folders up to date, skipping generation.');
-  }
+  execSync('node bin/sync-test-folders.js', { cwd: rootDir, stdio: 'inherit' });
 }
 
 if (process.env.SKIP_TGZ !== 'true') {
@@ -116,21 +97,6 @@ function hashDir(dir, h) {
       h.update(fs.readFileSync(full));
     }
   });
-}
-
-function hashTemplates(dir, h) {
-  fs.readdirSync(dir, { withFileTypes: true })
-    .filter(entry => entry.name !== 'node_modules' && !entry.name.startsWith('_v'))
-    .forEach(entry => {
-      const full = path.join(dir, entry.name);
-      if (entry.isDirectory()) {
-        hashTemplates(full, h);
-      } else if (entry.name === 'package.json.template' || entry.name === 'modes.json') {
-        h.update(fs.readFileSync(full, 'utf8'));
-      } else if (entry.name === 'test_base.js') {
-        h.update(full);
-      }
-    });
 }
 
 /**
